// src/gateway/protocol/schema/frames.d.ts
export declare const HelloOkSchema: Type.TObject<{
    type: Type.TLiteral<"hello-ok">;
    protocol: Type.TInteger;
    server: Type.TObject<{
        version: Type.TString;
        connId: Type.TString;
    }>;
    features: Type.TObject<{
        methods: Type.TArray<Type.TString>;
        events: Type.TArray<Type.TString>;
    }>;
    snapshot: Type.TObject<{
        presence: Type.TArray<Type.TObject<{
            host: Type.TOptional<Type.TString>;
            ip: Type.TOptional<Type.TString>;
            version: Type.TOptional<Type.TString>;
            platform: Type.TOptional<Type.TString>;
            deviceFamily: Type.TOptional<Type.TString>;
            modelIdentifier: Type.TOptional<Type.TString>;
            mode: Type.TOptional<Type.TString>;
            lastInputSeconds: Type.TOptional<Type.TInteger>;
            reason: Type.TOptional<Type.TString>;
            tags: Type.TOptional<Type.TArray<Type.TString>>;
            text: Type.TOptional<Type.TString>;
            ts: Type.TInteger;
            deviceId: Type.TOptional<Type.TString>;
            roles: Type.TOptional<Type.TArray<Type.TString>>;
            scopes: Type.TOptional<Type.TArray<Type.TString>>;
            instanceId: Type.TOptional<Type.TString>;
        }>>;
        health: Type.TAny;
        stateVersion: Type.TObject<{
            presence: Type.TInteger;
            health: Type.TInteger;
        }>;
        uptimeMs: Type.TInteger;
        configPath: Type.TOptional<Type.TString>;
        stateDir: Type.TOptional<Type.TString>;
        sessionDefaults: Type.TOptional<Type.TObject<{
            defaultAgentId: Type.TString;
            mainKey: Type.TString;
            mainSessionKey: Type.TString;
            scope: Type.TOptional<Type.TString>;
        }>>;
        authMode: Type.TOptional<Type.TUnion<[Type.TLiteral<"none">, Type.TLiteral<"token">, Type.TLiteral<"password">, Type.TLiteral<"trusted-proxy">]>>;
        updateAvailable: Type.TOptional<Type.TObject<{
            currentVersion: Type.TString;
            latestVersion: Type.TString;
            channel: Type.TString;
        }>>;
    }>;
    canvasHostUrl: Type.TOptional<Type.TString>;
    auth: Type.TOptional<Type.TObject<{
        deviceToken: Type.TOptional<Type.TString>;
        role: Type.TString;
        scopes: Type.TArray<Type.TString>;
        issuedAtMs: Type.TOptional<Type.TInteger>;
        deviceTokens: Type.TOptional<Type.TArray<Type.TObject<{
            deviceToken: Type.TString;
            role: Type.TString;
            scopes: Type.TArray<Type.TString>;
            issuedAtMs: Type.TInteger;
        }>>>;
    }>>;
    policy: Type.TObject<{
        maxPayload: Type.TInteger;
        maxBufferedBytes: Type.TInteger;
        tickIntervalMs: Type.TInteger;
    }>;
}>;